LayoutTests/fast/js/exception-thrown-from-new-expected.txt

This page tests exceptions thrown from 'new' expressions. If the test passes, you'll see a series of PASS messages below.

PASS: e1 should be 'TypeError: Result of expression 'f' [undefined] is not a constructor.' and is.
PASS: e2 should be 'TypeError: Result of expression 'g' [undefined] is not a function.' and is.
